Create Person_inlinefunction_constructors.cpp
Create a class Person which includes: character array name of size 64, age in numeric, character array address of size 64, and total salary in real numbers (divide salary in different components, if required). Make an array of objects of class Person of size 10.
(a) Write inline function that obtains the youngest and eldest age of a person from a class person using arrays.
(b) Write a program to develop the salary slip and display result by using constructors.