Create swap_friendfunction.cpp
a) Write a program to swap two numbers (create two classes) by using Friend function.
b) Write a program to create two classes DistM and DistF which store the values of distance. DistM stores distance in meters and centimetres and DistF stores distances in feet and inches. Read values for the class object and add one object of DistM with another object of DistF. Use a friend function for the addition operation and display answer in meter and centimetres.